refactor: merge global functionalities to DoobooContext
